+++ remix.config.js
@@ -11,9 +11,8 @@ module.exports = {
     "is-binary-path",
     "yargs-parser",
     "gitignore-parser",
-    "chalk",
   ],
-  // appDirectory: "app",
+  appDirectory: "src",
   // assetsBuildDirectory: "public/build",
   // serverBuildPath: "build/index.js",
   // publicPath: "/build/",
