+++ src/components/AuthorDistOther.tsx
