+++ src/components/Details.tsx
@@ -2,7 +2,7 @@ import { Spacer } from "./Spacer"
 import { makePercentResponsibilityDistribution } from "./Chart"
 import { Box, BoxTitle, CloseButton, DetailsKey, DetailsValue } from "./util"
 import { useOptions } from "../contexts/OptionsContext"
-import { HydratedGitBlobObject } from "../../parser/src/model"
+import { HydratedGitBlobObject } from "~/parser/model"
 import { dateFormatLong } from "../util"
 import styled from "styled-components"
 import { useState } from "react"
