+++ src/components/GlobalInfo.tsx
