+++ src/components/LegendFragment.tsx
