+++ src/components/LegendOther.tsx
