+++ src/extension-color.ts
