+++ src/lang-map.d.ts
