+++ src/routes/index.tsx
@@ -1,14 +1,14 @@
 import { json, LoaderFunction, useLoaderData } from "remix";
-import { ParserData } from "~/parser/src/model";
 import appStyles from "~/styles/App.css"
 import varsStyles from "~/styles/vars.css"
 import indexStyles from "~/styles/index.css"
 import chartStyles from "~/styles/Chart.css"
-import { parse } from "~/parser/src/parse.server";
-import { Providers } from "~/src/components/Providers";
-import { Container } from "~/src/components/util";
-import { SidePanel } from "~/src/components/SidePanel";
-import { Main } from "~/src/components/Main";
+import { Providers } from "~/components/Providers";
+import { Container } from "~/components/util";
+import { SidePanel } from "~/components/SidePanel";
+import { Main } from "~/components/Main";
+import { ParserData } from "~/parser/model";
+import { parse } from "~/parser/parse.server";
 
 export function links() {
   return [appStyles,
