+++ tsconfig.json
@@ -11,7 +11,7 @@
     "strict": true,
     "baseUrl": ".",
     "paths": {
-      "~/*": ["./app/*"]
+      "~/*": ["src/*"]
     },
 
     // Remix takes care of building everything in `remix build`.
